requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.1.1
djangorestframework==3.8.2
future==0.16.0
httpie==0.9.9
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pygments==2.2.0
python-http-client==3.1.0
pytz==2018.5
requests==2.19.1
sendgrid==5.6.0
serializers==0.2.4
uritemplate==3.0.0
urllib3==1.23
